Add multiple video download to HD videos
The user can now specify in the settings file what types of files should
be downloaded. The default is still the first trailer, though which file
this downloads might be different now than previously. Previously, the
script always looked for a file with "tlr1" in the name, now it
downloads the video for which the header "Trailer" is displayed to the
user.

This version does not support multiple files for the "web" video source,
which means that even HD videos will not properly respect the
video_types config if the script has to fall back to the "web" source
because HD iTunes files are not available.