Add command-line options
This commit adds command-line options for all config values. These
override the values specified in the config file.

The message to the user is currently incorrect if no config file is
found, but the user specifies options on the commadn line. In this case,
the script says it's using the defaults for all values.