Support the new Apple Trailers site structure
The Apple Trailers site dropped the separation of iTunes and Web files
and delivers more data via JSON files instead of HTML include files.
Previous to this commit, the download script could no longer download
480p files, since the Web trailers weren't supported anymore on the
Apple site.

This commit changes the script to only work with the newer JSON data,
which fixes support for 480p files.

Getting rid of HTML include file support means we can drop the BeautifulSoup
dependency. We also can stop reading the film title from the HTML page
title and instead pull it out of the JSON data. This saves a HTTP request for
each movie when a URL is given on the command line.