Rename functions and variables to snake case
Back when the library was started, it was written with camelCase
function and variable names. Now snake_case has become the clearly
preferred style, and since recent commits introduced breaking changes to
some central functions, we'll take this opportunity to clean up the
style and make it conform to PEP8.