Change unicode handling
Instead of trying to convert to unicode strings when dealing with the
video filenames, always handle them as unicode strings. Since we are now
getting the film titles from JSON, they always come out of the parser as
unicode strings. As long as we keep them that way, we shouldn't have any
problems reading or writing the download list file.