Support trailers marked "trailer 1"
For most trailers, the first trailer has the `title` attribute set to
`trailer`, but some of them have it set to `trailer 1`. Notably, the
Star Wars: The Rise of Skywalker trailer has this title. This commit
changes the logic to look for either of those titles when the
`video_types` setting is set to `single_trailer`.