Add download_all_urls option
This commit adds an option to supply a list of URLs for which the script
will download all files, regardless of other settings. This allows the
user to do something like only download Trailer #1 for most movies (with
the `video_types = single_trailer` setting), but for movies the user is
really interested in, to download all videos.