Move all imports to the top of the file
Originally some imports were done within functions in a misguided effort
to be more efficient when the script was imported into another script
instead of being run as a standalone program. However, pylint now
lints for that and the efficiency gains were probably imaginary anyway,
so we'll just move everything to the top now.