Handle extra spaces in video title
There has been at least one example of a trailer having the title
"Trailer " (with a trailing space) and this messed up the
`video_types` config logic when it was set to `single_trailer`. We now
strip extra whitespace from the beginning and end of the title, as well
as combining any duplicated spaces within the title to a single space.