Change behavior of single_trailer setting
We've seen at least one instance of a movie having a "Trailer 2", but no
"Trailer" or "Trailer 1". In this case, don't want the script to skip
the movie completely. Instead, we want it to download "Trailer 2".

Note: with the current logic, if such a movie gets a "Trailer 1" at a
later point in time, the script will also download that trailer.