Prevent repeat download in single_trailer mode
Previously there was a situation where the script would download
multiple trailers, even in single_trailer mode. If a movie only had a
"Trailer 2" and no "Trailer 1", the script would download "Trailer 2" in
single_trailer mode. Then, if a "Trailer 1" appeared for that movie, the
script would have downloaded that trailer as well.

This commit changes the behavior so that if the script has downloaded
any trailer for a particular movie, it will never download another
trailer in single_trailer mode.