Fix download detection when titles are munged
Recent changes introduced a bug where files were not being detected as
already downloaded in the `single_trailer` mode if the movie title had
characters that were removed when generating the filename. This commit
fixes that bug, so now it also correctly detects already-downloaded
trailers if the movie title contains invalid characters.