Disable pylint rules in specific versions only
Instead of putting rule disable comments in the code, where they have to
be understood by all versions of pylint, instead use command-line
arguments to disable rules for specific versions. This avoids the
situation where a rule is unknown in earlier versions of pylint, so the
disable comment generates and error.